Added modularity to in- and output
Created abstract classes for input- and output handler to allow for
choosing different input/output file types.
